story/NNO-13130 Sonar. Bug. "Cast one of the operands of this addition operation to a "long""

Closes NNO-13130

See merge request framework/n2o-framework!8734

+++ backend/n2o/n2o-mvc/src/main/java/net/n2oapp/framework/mvc/cache/LifetimeClientCacheTemplate.java
@@ -9,7 +9,7 @@ import java.util.Date;
  * Шаблон кэширования запросов по стратегии "Время жизни" (по умолчанию 1 минута).
  */
 public class LifetimeClientCacheTemplate extends ClientCacheTemplate {
-    private long lifetime = 1000 * 60;
+    private long lifetime = (long) 1000 * 60;
 
     public LifetimeClientCacheTemplate() {
     }
